SpaceX Application Implementation
This Application has been created using React. In this application there are 5 sections have been implemented in the landing page. Below have been discussed each section in detail

1) Header => First section is header where spacex logo is added on the left side.

2) Banner => This is the section where we have given tagline related to company and spaceX background image.

3) Search Form => This is the place from where we can search the rockets which we want to search. Here we have added basic validations as well if the form is empty then we will give alert to the user to fill the form first. After Filling the form when we click on search button we will get the data related to the query if there are no records then proper message will appear to the user.

4) Rockets List => This is the section where we have all the rockets which are available. When we click on these items Popup will appear with the complete details related to that selected rocket. Here we have pagination as well so when we select a particular page then according to the selected page data will receive and will be updated in the list.

5) Footer => This is the last section in the application where we have few details.

Here we are not dependent on API for data fetching. We are simply taking the data from json file. If we have an API then initially when the page loads then we should pass page = 1 and size = 10 (records per page) to the API and that API should return the data with one more detail called total records so that we can decide how many pages we can have at most. Later whenever we select specific page that page no and size need to pass to the API and the returned data need to update in the current list.

This website is responsive as well. I have attached below the Video URL for this website

https://watch.screencastify.com/v/7iuAWDhQVkUgzLzYzHaX